+++ src/utils/request.ts
@@ -16,17 +16,10 @@ export async function generateRequestData() {
   const opts = { ipfsGatewayURL: "https://ipfs.io" }; // can be your IFPS gateway if your work with ipfs schemas or empty object
   const ldCtx = (await getDocumentLoader(opts)(url)).document;
   const ldJSONStr = JSON.stringify(ldCtx);
-  // const ldBytes = byteEncoder.encode(ldJSONStr);
-  const typeId = await Path.getTypeIDFromContext(ldJSONStr, type);
-  const schemaHash = calculateCoreSchemaHash(byteEncoder.encode(typeId));
-  console.log("schemaId", schemaHash.bigInt().toString());
 
   // you can use custom IPFS
   const path = await Path.getContextPathKey(ldJSONStr, type, fieldName, opts);
   path.prepend([pathToCredentialSubject]);
-  const pathBigInt = await path.mtEntry();
-
-  console.log("path", pathBigInt.toString());
 
   // you can hash the value according to the datatype (that's how it is stored in core claim structure)
